+++ src/ladb_opencutlist/js/constants.js
@@ -1,7 +1,7 @@
 // CONSTANTS
 // ======================
 
-const EXTENSION_BUILD = '202503171618';   // /!\ Auto-generated line, do not edit //
+const EXTENSION_BUILD = '202503192005';   // /!\ Auto-generated line, do not edit //
 
 // UI /////